### 2023-03-13 #201 @whedon
[PRE REVIEW]: Planet_LB: Lattice-Boltzmann solutions for planetary geodynamics problems
**Submitting author:** @thecraigoneill (<a href="http://orcid.org/0000-0002-6034-1881">Craig O&#39;Neill</a>)
**Repository:** <a href="https://github.com/thecraigoneill/planet_LB" target ="_blank">https://github.com/thecraigoneill/planet_LB</a>
**Version:** v1.0.0

Due to the challenges of the COVID-19 pandemic, JOSE is currently operating in a "reduced service mode".
