### 2023-03-13 #201 @labarba
@thecraigoneill – your paper has a lot of issues with references, and also the figure needs size adjustment. Please make this fixes, and when done you can run the command `@whedon generate pdf` alone in a comment here.

To make intermediate tests of the paper compilation, you can use https://whedon.theoj.org
