### 2023-03-13 #201 @thecraigoneill
Potential reviewer is https://github.com/brmather
Ben has published with JOSS and is familiar with the process, and this code.
C
