+++ paper.bib
@@ -1,12 +1,13 @@
 
-@book{turcotte_geodynamics_2002,
+@book{Turcotte:2002,
 	title = {Geodynamics},
 	publisher = {Cambridge university press},
 	author = {Turcotte, DL and Schubert, G},
+	doi = {10.1017/CBO9780511807442},
 	year = {2002},
 }
 
-@book{mohamad_lattice_2019,
+@book{Mohamad:2019,
 	address = {London},
 	title = {Lattice {Boltzmann} {Method}: {Fundamentals} and {Engineering} {Applications} with {Computer} {Codes}},
 	isbn = {978-1-4471-7422-6 978-1-4471-7423-3},
@@ -21,7 +22,7 @@
 	file = {Submitted Version:/Users/craig/Zotero/storage/GSJ2Z5AV/Mohamad - 2019 - Lattice Boltzmann Method Fundamentals and Enginee.pdf:application/pdf},
 }
 
-@article{moresi_computational_2007,
+@article{Moresi:2007,
 	title = {Computational approaches to studying non-linear dynamics of the crust and mantle},
 	volume = {163},
 	issn = {00319201},
@@ -37,7 +38,7 @@
 	pages = {69--82},
 }
 
-@book{fowler_solid_2005,
+@book{Fowler:2005,
 	address = {Cambridge, UK},
 	edition = {Second edition},
 	title = {The solid earth: an introduction to global geophysics},
@@ -47,5 +48,6 @@
 	publisher = {Cambridge University Press},
 	author = {Fowler, C. M. R.},
 	year = {2005},
+	ISBN = {0-521-81066-3},
 	note = {OCLC: 817914139},
 }
