+++ paper.md
@@ -46,7 +46,7 @@ $$ f_i(r+c_i \Delta t, t + \Delta t) = f_i(r,t) + (\Delta t/{\tau})[f^{eq}_i(r,t
 
 The underlying grid is structured as shown in Figure 3, and the node numbering scheme is used both in the collision operator step (each node is weighted differently relative to the cell centre), and in the streaming step with advects the underlying distribution function, subject to boundary conditions. For full details of Lattice-Boltzmann implementation, see @Mohamad:2019.
 
-![Figure 3. Examples of LB grid structure and node numbering](D1Q2_D2Q9.png | width = 250)
+![Figure 3. Examples of LB grid structure and node numbering](D1Q2_D2Q9.png = 250x250)
 
 *Figure 3. Examples of LB grid structure and node numbering*
 
