### 2023-04-01 #201 @nicoguaro
@brmather, @lheagy, would you be interested in reviewing this submission for JOSE? It concerns a _Python package for 2D geodynamics using a lattice Boltzmann approach_.
