### 2023-04-16 #201 @nicoguaro
@jfpa, would you be interested in reviewing this submission for JOSE? It concerns a _Python package for 2D geodynamics using a lattice Boltzmann approach_.
