### 2023-04-17 #205 @whedon
[REVIEW]: Planet_LB: Lattice-Boltzmann solutions for planetary geodynamics problems
**Submitting author:** @thecraigoneill (<a href="http://orcid.org/0000-0002-6034-1881">Craig O&#39;Neill</a>)
**Repository:** https://github.com/thecraigoneill/planet_LB
**Branch with paper.md** (empty if default branch):
**Version:** V1.1

- [x] #214

- [x] #213

- [x] #218
- [x] #219

- [x] #215
- [x] #216

- [x] #217
