### 2023-04-17 #205 @nicoguaro
@brmather, @jfpa, this is the space where the review process takes form. Please see the checklist for each one above and tick the boxes when you see that the criterion is satisfied.

I will be here to answer the questions that you might have.

Let us use as a tentative timeframe the third week of May, is that OK for you?
