### 2023-05-29 #205 @jfpa
Dear Dr. Guarín-Zapata,

I'm making progress toward completing the review. I hope to have it ready
by the end of the week.

Cheers,
Pani

On Mon, May 29, 2023 at 9:07 AM Nicolás Guarín-Zapata <
***@***.***> wrote:

> @brmather <https://github.com/brmather>, @jfpa <https://github.com/jfpa>,
> would you mind giving us an update on how your review is going? Is there
> something that @thecraigoneill <https://github.com/thecraigoneill> or I
> can do to help you move forward?
>
