### 2023-06-09 #205 @jfpa
Dear Prof. Guarin @nicoguaro, I'm unable to complete the checklist (the boxes are deactivated). Could you please indicate a way to proceed?
