### 2023-06-12 #205 @brmather
Dear @nicoguaro,

I have reviewed the software and found that it is looking spiffy. The Lattice Boltzmann solver works very efficiently and the examples are, indeed, very useful to Earth Science applications. My only issue with the repository is that the installation instructions are not sufficiently outlined in the README. It needs to have a separate subheading titled "Installation" and the dependencies need to be listed.

#### Minor comments

- I was initially confused that there is no `setup.py` file included in the repository, yet installing `planet_LB` via pip is supported. I suppose that means the package is manually created by the author and pushed to PyPI. If so, there is an opportunity to set up a GitHub action to handle this package deployment whenever a new release is published.
- The tests within `planet_LB` are _not_ exhaustive to say the least. Maybe a few more tests could be included in there to benchmark numerical solutions against analytical solutions. FYI, GitHub actions can test software on a range of python versions and system architectures.
- Community guidelines for contributions need to be included.
