### 2023-06-13 #205 @nicoguaro
@jfpa, the problem checking te boxes persists or are you able to check them now?
