### 2023-06-13 #205 @jfpa
> @jfpa, the problem checking te boxes persists or are you able to check them now?

No, I'm not able to check them.
