### 2023-06-14 #205 @jfpa
Dear colleagues, Prof. Barba @labarba, Prof. Guarín-Zapata @nicoguaro et al.,

Here are some "issues" related to the checklist above. Thanks for the opportunity of reviewing this submission. Please, let me know if this formatting works or if there is a different way of presenting the issues.

**Functionality/Performance**
I suggest either optimizing the examples for use in the classroom or stating that they're intended for asynchronous activities.

For example, the ‘Lid-driven flow example’ (‘LB_id_flow.ipynb’) is relatively inefficient in the calculation of the cell [3]. For 40,000 steps it takes too long to calculate the benchmark case in my laptop (likely the way a student would use it in class). If these cases are intended for use in the classroom, I’d suggest using a different approach to explore Mohamad's (2011) example. Currently, it is ~120 minutes for 40,000 steps. However, a cool perk is that you can visualize the cavity flow velocity for any quantity of steps. In this vein, would it be useful to save the steps to create a visualization of the evolution?

**Documentation**
I suggest including the installation requirements (packages) in the README to make the test easier, explicitly stating that those are the needed packages (i.e., expanding on the ‘requirements.txt’ file).

**Software paper**
I suggest expanding the references on the implementation of Lattice-Boltzmann modeling to geodynamics and providing key references beyond Mohamad (2011), intended for undergraduate students. For example, Chen & Doolen (1998), Dong et al. (2008), Guo et al. (2000), and D’Humieres (1992).

**Suggested references**
Chen, S., & Doolen, G. D. (1998). Lattice Boltzmann Method for Fluid Flows. Annual Reviews of Fluid Mechanics, 30, 329–364.

D’Humieres, D. (1992). Generalized Lattice-Boltzmann Equations. In B. D. Shizgal & D. P. Weaver (Eds.), Rarefied Gas Dynamics - Theory and Simulations (Vol. 159, pp. 450–458). Progress in Astronautics and Aeronautics.

Dong, Y.-H., Sagaut, P., & Marie, S. (2008). Inertial consistent subgrid model for large-eddy simulation based on the lattice Boltzmann method. Physics of Fluids, 20(035104), 1–13. https://doi.org/10.1063/1.2842379.

Guo, Z., Shi, B., & Wang, N. (2000). Lattice BGK Model for Incompressible Navier-Stokes Equation. Journal of Computational Physics, 165, 288–306. https://doi.org/10.1006/jcph.2000.6616.
