### 2023-06-20 #205 @nicoguaro
@thecraigoneill, both reviewers have marked the checklist for your submission. Nevertheless, they left some comments related to "minor issues". Can you address them?
