+++ paper.bib
@@ -49,3 +49,61 @@
 	year = {2005},
 	note = {OCLC: 817914139},
 }
+
+@article{chen_lattice_1998,
+	title = {Lattice {Boltzmann} {Method} for {Fluid} {Flow}},
+	volume = {30},
+	issn = {0066-4189, 1545-4479},
+	url = {https://www.annualreviews.org/doi/10.1146/annurev.fluid.30.1.329},
+	doi = {10.1146/annurev.fluid.30.1.329},
+	language = {en},
+	number = {1},
+	urldate = {2023-06-26},
+	journal = {Annual Review of Fluid Mechanics},
+	author = {Chen, Shiyi and Doolen, Gary D.},
+	month = jan,
+	year = {1998},
+	pages = {329--364},
+}
+
+@article{dong_inertial_2008,
+	title = {Inertial consistent subgrid model for large-eddy simulation based on the lattice {Boltzmann} method},
+	volume = {20},
+	issn = {1070-6631, 1089-7666},
+	url = {https://pubs.aip.org/aip/pof/article/865171},
+	doi = {10.1063/1.2842379},
+	language = {en},
+	number = {3},
+	urldate = {2023-06-26},
+	journal = {Physics of Fluids},
+	author = {Dong, Yu-Hong and Sagaut, Pierre and Marie, Simon},
+	month = mar,
+	year = {2008},
+	pages = {035104},
+}
+
+@article{guo_lattice_2000,
+	title = {Lattice {BGK} {Model} for {Incompressible} {Navier}–{Stokes} {Equation}},
+	volume = {165},
+	issn = {00219991},
+	url = {https://linkinghub.elsevier.com/retrieve/pii/S0021999100966166},
+	doi = {10.1006/jcph.2000.6616},
+	language = {en},
+	number = {1},
+	urldate = {2023-06-26},
+	journal = {Journal of Computational Physics},
+	author = {Guo, Zhaoli and Shi, Baochang and Wang, Nengchao},
+	month = nov,
+	year = {2000},
+	pages = {288--306},
+}
+
+@incollection{dhumieres_generalized_1992,
+	series = {Progress in {Astronautics} and {Aeronautics}},
+	title = {Generalized {Lattice} {Boltzmann} {Equations}},
+	volume = {159},
+	booktitle = {Rarefied {Gas} {Dynamics}: {Theory} and {Simulations}},
+	author = {d’Humieres, D},
+	year = {1992},
+	pages = {450--458},
+}
