+++ setup.py
@@ -1,0 +1,9 @@
+import setuptools
+
+setuptools.setup(
+        name="planet_LB",
+        version="1.0.0",
+        author="Craig O'Neill",
+        description="Lattice Boltzmann Solver for python",
+        packages=["planet_LB"]
+)
