+++ planet_LB/planet_LB.py
@@ -33,7 +33,7 @@ class planet_LB():
         f2[m] = f2[m-1]
     return T
 
-  def init_1D_LB(self,dt,dx,mstep,m,Tinit,TSurf,H):
+  def init_1D_LB(self,dt,dx,mstep,m,Tinit,TSurf,H,Qbase,K,k):
     f1 = np.array([])
     f2 = np.array([])
     feq = np.array([])
@@ -41,12 +41,12 @@ class planet_LB():
     scale=(60*60*24*365.25*1e6)
     x = np.arange(0,dx*(m+1),dx)
     ck=dx/dt
-    csq=ck*ck
-    K=1e-6
+    cksq=ck*ck
+    #K=1e-6
     #k=3
     #flux=80e-3
-
-    omega=1.0/(K/(dt*csq)+0.5) 
+    csq=1/np.sqrt(2)
+    omega=1.0/(K/(dt*cksq)+0.5) 
     T = np.ones((m+1))*Tinit
     f1 = 0.5*T
     f2 = 0.5*T
@@ -63,13 +63,50 @@ class planet_LB():
       'f2':f2,
       'feq':feq,
       'x':x,
-      'T':T
+      'Qbase':Qbase,
+      'K':K,
+      'k':k,
+      'dx':dx,
+      'csq':csq,
     }
-
-    #T = self.LB1D(LB_struct)
-
-    #return(T,x)
     return(LB_struct)
+
+  def LB1D_Qbase(self,LB_struct):
+    mstep = LB_struct['mstep']
+    m = LB_struct['m']
+    dt = LB_struct['dt']
+    dx = LB_struct['dx']
+    f1  = LB_struct['f1']
+    f2 = LB_struct['f2']
+    H = LB_struct['H']
+    K=LB_struct['K']
+    omega = LB_struct['omega']
+    TSurf = LB_struct['TSurf']
+    Qbase = LB_struct['Qbase']
+    k = LB_struct['k']
+    #csq = LB_struct['csq']
+    c=dx/dt
+    csq = LB_struct['csq']
+
+    #print("Flux scale 1:",Qbase/(k*omega),csq)
+    #print("Flux scale 2:",Qbase*dx/k)
+    timestep = dt*np.arange(1,mstep+1,1)
+    n = np.arange(0,m+1,1)
+    for kk in timestep:
+        #Collision step
+        T = f1 + f2
+        feq = 0.5*T
+        # cs^2 = 0.5
+        f1 = (1.0 - omega)*f1 + omega*feq + dt*0.5*H   # H is heat production
+        f2 = (1.0 - omega)*f2 + omega*feq + dt*0.5*H   # Note 
+        #Streaming step
+        f1 = np.roll(f1,1)
+        f2=np.roll(f2,-1)
+        #Boundary condition
+        f1[0] = TSurf - f2[0]   #constant temperature wall at x=0
+        f2[m] =  f1[m-1] + f2[m-1] - f1[m] + Qbase*dx/(k)        # von neumann boundary condition, x=L,
+    return T
+
 
   #@jit(nopython=True)
   def loopy_loop(self,n,m,vx,vy,cx,cy,rho,w):
[…]
