+++ README.md
@@ -10,6 +10,9 @@ The preferred procedure for installation is using pip as follows:
 
 pip install planet_LB
 
+You can install from the git repository download as well, using the command
+pip install .
+from the main directory.
 
 
 # Requirements
