+++ paper.bib
@@ -52,7 +52,7 @@
 	note = {OCLC: 817914139},
 }
 
-@article{chen_lattice_1998,
+@article{chen:1998,
 	title = {Lattice {Boltzmann} {Method} for {Fluid} {Flow}},
 	volume = {30},
 	issn = {0066-4189, 1545-4479},
@@ -68,7 +68,7 @@
 	pages = {329--364},
 }
 
-@article{dong_inertial_2008,
+@article{dong:2008,
 	title = {Inertial consistent subgrid model for large-eddy simulation based on the lattice {Boltzmann} method},
 	volume = {20},
 	issn = {1070-6631, 1089-7666},
@@ -84,7 +84,7 @@
 	pages = {035104},
 }
 
-@article{guo_lattice_2000,
+@article{guo:2000,
 	title = {Lattice {BGK} {Model} for {Incompressible} {Navier}–{Stokes} {Equation}},
 	volume = {165},
 	issn = {00219991},
@@ -100,7 +100,7 @@
 	pages = {288--306},
 }
 
-@incollection{dhumieres_generalized_1992,
+@incollection{dhumieres:1992,
 	series = {Progress in {Astronautics} and {Aeronautics}},
 	title = {Generalized {Lattice} {Boltzmann} {Equations}},
 	volume = {159},
