### 2023-08-01 #205 @thecraigoneill
@nicoguaro I've completed the modifications requested by the reviewers. I'm attaching my comments regarding what was done below:

Comments from @brmather:

_My only issue with the repository is that the installation instructions are not sufficiently outlined in the README. It needs to have a separate subheading titled "Installation" and the dependencies need to be listed._

Both of these requests are done.

_I was initially confused that there is no setup.py file included in the repository, yet installing planet_LB via pip is supported. I suppose that means the package is manually created by the author and pushed to PyPI. If so, there is an opportunity to set up a GitHub action to handle this package deployment whenever a new release is published._

This is a good point. I’ve included the setup file in the repository now, but haven’t automated the pip push call, as I’m a little more conservative with that distribution than github, but this is certainly a possibility once the code base is more mature.

_The tests within planet_LB are not exhaustive to say the least. Maybe a few more tests could be included in there to benchmark numerical solutions against analytical solutions. FYI, GitHub actions can test software on a range of python versions and system architectures._

There are two things here – the benchmarks themselves, and the commit tests. I think the benchmarks test the core functionality of the code fairly well, though, for time-dependent problems, they can take a while to run. I’ve added some more geothermal testing in a new ipynb as well. Regarding implementing the commit tests (in tests.py): The benchmarks generally run through time and so can make for a slow commit if the full suite are run. I’ve opted to test the basic core functionality instead. I’ve updated with some more tests to make sure the 2D structures are behaving each commit.

_Community guidelines for contributions need to be included._
Community guidelines included!

Comments from @jfpa

I _suggest either optimizing the examples for use in the classroom or stating that they're intended for asynchronous activities._

Good point. Some of these are for robust benchmarking, and probably aren’t best suited for the classroom. I’ve updated the ipynb to note this, particularly for the lid problem.

_However, a cool perk is that you can visualize the cavity flow velocity for any quantity of steps. In this vein, would it be useful to save the steps to create a visualization of the evolution?_
Absolutely! I had the functionality hidden away in there. I’ve now made this more explicit, with its own section and a revised codeblock in the ipynb.

_I suggest including the installation requirements (packages) in the README to make the test easier, explicitly stating that those are the needed packages (i.e., expanding on the ‘requirements.txt’ file)._
This has been done.

_I suggest expanding the references on the implementation of Lattice-Boltzmann modeling to geodynamics and providing key references beyond Mohamad (2011), intended for undergraduate students. For example, Chen & Doolen (1998), Dong et al. (2008), Guo et al. (2000), and D’Humieres (1992)._

Added the suggested references.
