+++ paper.bib
@@ -100,12 +100,14 @@
 	pages = {288--306},
 }
 
-@incollection{dhumieres:1992,
-	series = {Progress in {Astronautics} and {Aeronautics}},
+@inbook{dhumieres:1992,
 	title = {Generalized {Lattice} {Boltzmann} {Equations}},
 	volume = {159},
 	booktitle = {Rarefied {Gas} {Dynamics}: {Theory} and {Simulations}},
 	author = {d’Humieres, D},
 	year = {1992},
 	pages = {450--458},
+	doi = {10.2514/5.9781600866319.0450.0458},
+	url = {https://arc.aiaa.org/doi/10.2514/5.9781600866319.0450.0458},
+	urldate = {2023-08-01},
 }
