### 2023-08-01 #205 @thecraigoneill
@nicoguaro the paper pdf in the github repo is now up to date too (as you no doubt saw from my preceding 10 comment wall-of-shame editing effort!). I think this takes care of the revisions on my end. I'll await your feedback before thinking about zenodo or somesuch for DOIs.
cheers
Craig
