### 2023-08-01 #205 @nicoguaro
@jfpa and @brmather, did @thecraigoneill address your concerns regarding ``planet_LB``?

If so, and seeing that you both already marked all the boxes, would you recommend the work for publication?
