### 2023-08-02 #205 @nicoguaro
@thecraigoneill, I have created a [PR](https://github.com/thecraigoneill/planet_LB/pull/3) with some style suggestions. Also, I think that you can change the bib file to get the right capitalization for *Cambridge University Press*.

Regarding the figures in the paper, I have some questions/suggestions:

- You are using a colormap that is divergent (I think that it is RdGY). These colormaps are commonly used when one has a particular value that want to highlight and present results above/below it. In the case of figure 1, this seems to be the value of 400. Is there anything special about 400 there?

- I suggest that the figures are explicitly mentioned in the text of the paper.

- The labels for the vertical axes of the figures are present but the horizontal ones are missing. Also, what are the units?

- I am guessing that the color represents temperature (T) but it is not explicitly stated in the caption. It is also missing the units.
