+++ paper.bib
@@ -1,7 +1,7 @@
 
 @book{Turcotte:2002,
 	title = {Geodynamics},
-	publisher = {Cambridge university press},
+	publisher = {Cambridge University Press},
 	author = {Turcotte, DL and Schubert, G},
 	doi = {10.1017/CBO9780511807442},
 	year = {2002},
