### 2023-08-03 #205 @thecraigoneill
Hi @nicoguaro

_I have created a https://github.com/thecraigoneill/planet_LB/pull/3 with some style suggestions. Also, I think that you can change the bib file to get the right capitalization for Cambridge University Press._
I've updated the reference.

_Regarding the figures in the paper, I have some questions/suggestions:

    You are using a colormap that is divergent (I think that it is RdGY). These colormaps are commonly used when one has a particular value that want to highlight and present results above/below it. In the case of figure 1, this seems to be the value of 400. Is there anything special about 400 there?_

The base is at 800C, so 400C is halfway point. In convection literature/geodynamics (my background) it's pretty common to use diverging colorpalets as they show the thermal boundary layers well, and the RdGY is a nice modern version.

    _I suggest that the figures are explicitly mentioned in the text of the paper._
I've dropped mention in the relevant places.

    _The labels for the vertical axes of the figures are present but the horizontal ones are missing. Also, what are the units?_
Added units/labels to update figures.

    _I am guessing that the color represents temperature (T) but it is not explicitly stated in the caption. It is also missing the units._
On top of the Figure changes, I updated the captions to reflect this information.

Ok - I think that is it?
cheers
C
