### 2023-08-07 #205 @thecraigoneill
Hi @nicoguaro
Created a new release as requested, its tab is V1.1. Linked here:
[https://github.com/thecraigoneill/planet_LB/releases/tag/V1.1](https://github.com/thecraigoneill/planet_LB/releases/tag/V1.1)
