### 2023-08-09 #205 @labarba
hi @thecraigoneill — please check your proof for final adjustments. I noticed that the caption under Figures 1 and 2 are repeated in the text, with italics. Also, in parentheticals, replace the string "eg." with "e.g.," (including the comma). For Figure 3 move the caption text into an actual caption.
