### 2023-08-10 #205 @thecraigoneill
@labarba @nicoguaro - thanks guys. Manuscript has been cross-checked for all "eg". Captions inserted into each Figure instance (and removed from text). The final pdf generated looks.... reasonable (?!) to me, I don't see any more grammatical or formatting issues, so I think that's it?
